auth wbauth rename
