+++ wandb/sdk/lib/wbauth/authenticate.py
